requirements.txt
PyQt6
PyPubSub
qasync